Add new shortcode for External Link With target='_blank' in Hugo Markdown
https://adityatelange.in/blog/hugo-external-link-new-tab/

Implementaion
Hugo offers a way to render using a predefined template to Markdown(.md) with the help of Shortcodes. We can create such a shortcode for *Creating an Exeternal link with with target='_blank'.

Create a folder in /layouts named shortcodes/

Create a file named newtabref.html which will act as a templete for our shortcode :=

layouts/
├── _default/
├── partials/
└── shortcodes/
    └── newtabref.html <- new file created
In newtabref.html put the following code:
<a href="{{ .Get "href" }}" rel="noopener" target="_blank">{{ .Get "title" }}</a>
Now you can simply use
{{< newtabref  href="<YOUR EXTERNAL LINk>" title="<TITLE FOR LINK>" >}}
Thus we have created an External Link With target=’_blank’ In Hugo Markdown !!! :D